requirements.txt
python.exe -m pip install --upgrade pip
requests
wheel
lxml
pycryptodome
fake_useragent
